+++ examples/openapi/auth-apikey/src/main.rs
@@ -48,13 +48,11 @@ async fn main() -> Result<(), std::io::Error> {
     }
     tracing_subscriber::fmt::init();
 
-    let listener = TcpListener::bind("127.0.0.1:3000");
     let api_service =
         OpenApiService::new(Api, "Authorization Demo", "1.0").server("http://localhost:3000/api");
     let ui = api_service.swagger_ui();
 
-    poem::Server::new(listener)
-        .await?
+    poem::Server::new(TcpListener::bind("127.0.0.1:3000"))
         .run(Route::new().nest("/api", api_service).nest("/", ui))
         .await
 }