+++ examples/openapi/oneof/src/main.rs
@@ -1,4 +1,4 @@
-use poem::{listener::TcpListener, Route};
+use poem::{listener::TcpListener, Route, Server};
 use poem_openapi::{payload::Json, Object, OneOf, OpenApi, OpenApiService};
 
 #[derive(Object, Debug, PartialEq)]
@@ -36,12 +36,10 @@ async fn main() -> Result<(), std::io::Error> {
     }
     tracing_subscriber::fmt::init();
 
-    let listener = TcpListener::bind("127.0.0.1:3000");
     let api_service = OpenApiService::new(Api, "Oneof", "1.0").server("http://localhost:3000/api");
     let ui = api_service.swagger_ui();
 
-    poem::Server::new(listener)
-        .await?
+    Server::new(TcpListener::bind("127.0.0.1:3000"))
         .run(Route::new().nest("/api", api_service).nest("/", ui))
         .await
 }