+++ examples/openapi/poem-extractor/src/main.rs
@@ -1,4 +1,4 @@
-use poem::{listener::TcpListener, web::Data, EndpointExt, Route};
+use poem::{listener::TcpListener, web::Data, EndpointExt, Route, Server};
 use poem_openapi::{payload::PlainText, OpenApi, OpenApiService, PoemExtractor};
 
 struct Api;
@@ -18,13 +18,11 @@ async fn main() -> Result<(), std::io::Error> {
     }
     tracing_subscriber::fmt::init();
 
-    let listener = TcpListener::bind("127.0.0.1:3000");
     let api_service =
         OpenApiService::new(Api, "Poem Extractor", "1.0").server("http://localhost:3000/api");
     let ui = api_service.swagger_ui();
 
-    poem::Server::new(listener)
-        .await?
+    Server::new(TcpListener::bind("127.0.0.1:3000"))
         .run(
             Route::new()
                 .nest("/api", api_service.data(100i32))