+++ examples/openapi/upload/src/main.rs
@@ -1,6 +1,6 @@
 use std::collections::HashMap;
 
-use poem::{error::BadRequest, listener::TcpListener, Result, Route};
+use poem::{error::BadRequest, listener::TcpListener, Result, Route, Server};
 use poem_openapi::{
     param::Path,
     payload::{Binary, Json},
@@ -90,7 +90,6 @@ async fn main() -> Result<(), std::io::Error> {
     }
     tracing_subscriber::fmt::init();
 
-    let listener = TcpListener::bind("127.0.0.1:3000");
     let api_service = OpenApiService::new(
         Api {
             status: Mutex::new(Status {
@@ -104,8 +103,7 @@ async fn main() -> Result<(), std::io::Error> {
     .server("http://localhost:3000/api");
     let ui = api_service.swagger_ui();
 
-    poem::Server::new(listener)
-        .await?
+    Server::new(TcpListener::bind("127.0.0.1:3000"))
         .run(Route::new().nest("/api", api_service).nest("/", ui))
         .await
 }