+++ examples/poem/nested-routing/src/main.rs
@@ -17,6 +17,7 @@ async fn main() -> Result<(), std::io::Error> {
     tracing_subscriber::fmt::init();
 
     let app = Route::new().nest("/api", api());
-    let server = Server::new(TcpListener::bind("127.0.0.1:3000")).await?;
-    server.run(app).await
+    Server::new(TcpListener::bind("127.0.0.1:3000"))
+        .run(app)
+        .await
 }