+++ examples/poem/tonic/src/main.rs
@@ -37,7 +37,7 @@ async fn main() -> Result<(), std::io::Error> {
         format!("/{}", GreeterServer::<MyGreeter>::NAME),
         GreeterServer::new(MyGreeter).compat(),
     );
-    let listener = TcpListener::bind("127.0.0.1:3000");
-    let server = Server::new(listener).await?;
-    server.run(app).await
+    Server::new(TcpListener::bind("127.0.0.1:3000"))
+        .run(app)
+        .await
 }