+++ examples/poem/tower-layer/src/main.rs
@@ -21,7 +21,7 @@ async fn main() -> Result<(), std::io::Error> {
         "/",
         get(hello).with(RateLimitLayer::new(5, Duration::from_secs(30)).compat()),
     );
-    let listener = TcpListener::bind("127.0.0.1:3000");
-    let server = Server::new(listener).await?;
-    server.run(app).await
+    Server::new(TcpListener::bind("127.0.0.1:3000"))
+        .run(app)
+        .await
 }