+++ poem/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# Unreleased
+
+- Use percent-encoding before adding cookies to the header.
+
 # [1.0.26] 2021-11-15
 
 - Fix Cors middleware response incorrect `Access-Control-Allow-Headers` header.